+++ src/lib/Components/UniversalSelect.svelte
@@ -40,6 +40,7 @@
 	</div>
 {:else}
 	<Select
+		readonly
 		options={items}
 		placeholder={$lang('options')}
 		value={selected}