+++ src/lib/Main/Camera/HLS.svelte
@@ -132,7 +132,7 @@
 		{muted}
 		{controls}
 		style:display={$editMode ? 'none' : 'initial'}
-		style:width={responsive ? '100%' : 'calc(14.5rem * 2 + 0.4rem)'}
+		style:max-width={responsive ? '100%' : 'calc(14.5rem * 2 + 0.4rem)'}
 		style:object-fit={size}
 	></video>
 {/if}