+++ CHANGELOG.md
@@ -1,17 +1,46 @@
 # Changelog FUSION pro
 
+## [0.0.3] - 2024-06-16
+
+### Added
+
+- Read-only selection. Disable search input in mode selection
+- LightModal: support effects
+
+### Remove
+
+- Select: navigation by keyboard
+- Modal: close by drag
+
+### Fixed
+
+- Camera: max width
+- WeatherForecast: wrap days
+- UpdateModal: fix style and hide buttons if installed latest
+- LightModal: replace the power switch with a button switch
+- Horizontal stack: responsive
+- Modal
+  - Touch scroll
+  - Action buttons for mobile view
+  - Less empty spaces for mobile view
+  - Sticky action button in edit mode
+  - Styled scrollbar
+
 ## [0.0.2] - 2024-06-15
 
 ### Added
+
 - UniversalSelect component: show icon buttons or dropdown list when states have more than six items.
 - Vacuum: Fan speed modes
 - Climate: WheelPicker replaces to SlidePicker component for set temperature
 - WheelPicker component: live update
 
 ### Remove
+
 - Climate: WheelPicker component for set temperature
 
 ### Fixed
+
 - Vacuum: State and battery display in one line with icon
 - WheelPicker component: bugs
 - Small style fixes